#9 [FEAT] : 피드백 문장 color span 재활용
