+++ app/src/main/java/com/talkable/presentation/feedback/FeedbackExpressionFragment.kt
@@ -1,17 +1,14 @@
 package com.talkable.presentation.feedback
 
 import android.content.res.Resources
-import android.text.SpannableString
-import android.text.Spanned
-import android.text.style.ForegroundColorSpan
 import android.view.View
 import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.talkable.R
 import com.talkable.core.base.BindingFragment
 import com.talkable.core.util.context.pxToDp
-import com.talkable.core.util.fragment.colorOf
 import com.talkable.core.util.fragment.statusBarColorOf
 import com.talkable.databinding.FragmentTalkFeedbackExpressionBinding
+import com.talkable.presentation.FeedbackTextColor
 import com.talkable.presentation.feedback.model.FeedbackExpressionModel
 import com.talkable.presentation.feedback.model.FeedbackGrammarModel
 import com.talkable.presentation.talk.feedback.TalkFeedbackLearnedAdapter
@@ -50,21 +47,8 @@ class FeedbackExpressionFragment :
     }
 
     private fun setAfterAnswerTextColor(fullText: String, partsText: List<String>) {
-        val spannableString = SpannableString(fullText)
-
-        partsText.forEach { part ->
-            val startIndex = fullText.indexOf(part)
-            if (startIndex != -1) {
-                val endIndex = startIndex + part.length
-                val span = ForegroundColorSpan(colorOf(R.color.main_4))
-                spannableString.setSpan(
-                    span,
-                    startIndex,
-                    endIndex,
-                    Spanned.SPAN_EXCLUSIVE_EXCLUSIVE
-                )
-            }
-        }
+        val spannableString =
+            FeedbackTextColor(requireContext()).setAfterAnswerTextColor(fullText, partsText)
 
         binding.tvTalkFeedbackAfter.text = spannableString
     }
